No longer try to detect spelling errors outside of editable content.
